requirements.txt
fastapi
numpy
Pillow
pydantic
tensorflow
uvicorn
gunicorn